smooth normals

+++ include/terrain.hpp
@@ -4,7 +4,6 @@
 #include <cstdint>
 #include <webgpu/webgpu.hpp>
 #include "glm/ext/vector_float3.hpp"
-#include "glm/ext/vector_int3.hpp"
 #include "renderer/buffer.hpp"
 #include "renderer/layout.hpp"
 #include "renderer/program.hpp"
@@ -27,7 +26,7 @@ class Terrain {
   void draw(wgpu::TextureView view, wgpu::CommandEncoder encoder, wgpu::BindGroup cameraBindGroup);
 
  private:
-  BufferGroup<glm::ivec3, std::array<std::int32_t, 16>, std::array<std::uint32_t, 2>> m_tables;
+  BufferGroup<glm::vec3, std::array<std::int32_t, 16>, std::array<std::uint32_t, 2>> m_tables;
   VertexBuffer<TerrainVertex> m_vertices;
   ComputeProgram m_generator;
   RenderProgram m_program;
